added details button, still need positioning
