+++ sidebar/panel.css
@@ -2,6 +2,7 @@ body {
   height: 95%;
   font: caption;
   margin: 0;
+  margin-top: 0.5em;
   padding: 0.25em;
   background-color: white;
   overflow-y: scroll;
@@ -31,6 +32,13 @@ header .right {
   top: 0.25em;
 }
 
+header button#back-button:before {
+  content: url("data:image/svg+xml, %3Csvg xmlns='http://www.w3.org/2000/svg' class='down' width='12' height='12' viewBox='0 0 12 12'%3E%3Cpolygon points='1 6, 11 1, 11 11' fill='currentColor' /%3E%3C/svg%3E");
+  position: relative;
+  top:  1px;
+  padding-right:  4px;
+}
+
 main .view {
   display:  none;
 }
@@ -58,9 +66,8 @@ footer .info {
   margin: 0;
   padding: 0;
   display: grid;
-  grid-auto-columns: 80px 1fr;
+  grid-auto-columns: 70px 1fr;
   grid-template-rows: auto;
-  font-size: 125%;
 }
 
 footer .info .label {
@@ -84,19 +91,19 @@ footer .info .value {
 }
 
 footer .footer-buttons {
+  margin-top: 1rem;
+  margin-left: 0.5em;
+  margin-right: 0.5em;
   padding-top: 10px;
   display: flex;
-  justify-content: space-evenly;
-  align-content: space-between;
+  justify-content: space-between;
   flex-flow: row wrap;
   row-gap: 0.5rem;
-  margin-top: 1.25rem;
 }
 
 footer button {
-  padding-left: 1em;
-  padding-right: 1em;
   text-align: center;
+  width: 8em;
 }
 
 button:focus {
