+++ sidebar/panel.html
@@ -25,11 +25,6 @@
   </main>
 
   <footer>
-    <div id="info-ruleset" class="info">
-      <span class="label"></span>
-      <span class="value"></span>
-    </div>
-
     <div id="info-title" class="info">
       <span class="label"></span>
       <span class="value"></span>
@@ -40,12 +35,14 @@
       <span class="value"></span>
     </div>
 
+    <div id="info-ruleset" class="info">
+      <span class="label"></span>
+      <span class="value"></span>
+    </div>
+
     <div class="footer-buttons">
-      <div class="filler">&nbsp;</div>
       <button id="export-button">Export</button>
-      <div class="filler">&nbsp;</div>
       <rerun-evaluation-button></rerun-evaluation-button>
-      <div class="filler">&nbsp;</div>
       <button id="preferences-button"></button>
     </div>
