+++ sidebar/resultSummary.css
@@ -3,6 +3,7 @@
 table {
   margin-right: auto;
   margin-left: auto;
+  margin-bottom: 1.5em;
 }
 
 th, td {
