+++ sidebar/viewRuleResult.js
@@ -1,9 +1,9 @@
 /* viewRuleResult.js */
 
 const getMessage = browser.i18n.getMessage;
-const basicProps = ['result', 'tagName', 'position', 'role', 'actionMessage'];
 
 import { getOptions } from '../storage.js';
+import ViewRuleResultCSV  from './viewRuleResultCSV.js';
 
 export default class ViewRuleResult {
   constructor(id, handleRowSelection) {
@@ -45,152 +45,12 @@ export default class ViewRuleResult {
     this.elementResults = {};
 
     this.initGrid();
-  }
 
-  // Not all element information objects
-  // will define accessible names, descriptions
-  // or error descriptions
-  getAccNameProps () {
-    let props = [];
-    for (let er in this.elementResults) {
-      let info = this.elementResults[er];
-      if (info.accNameInfo) {
-        if (info.accNameInfo.name) {
-          if (props.indexOf('name') < 0) {
-            props.push('name');
-            props.push('name_source');
-          }
-        }
-
-        if (info.accNameInfo.desc) {
-          if (props.indexOf('desc') < 0) {
-            props.push('desc');
-            props.push('desc_source');
-          }
-        }
-
-        if (info.accNameInfo.error_desc) {
-          if (props.indexOf('error_desc') < 0) {
-            props.push('error_desc');
-            props.push('error_desc_source');
-          }
-        }
-      }
-    }
-    return props;
-  }
-
-  // If one element information object has
-  // CCR information all element information
-  // objects will
-  getVisibilityProps () {
-    let props = [];
-    if (this.elementResults) {
-      let id = Object.keys(this.elementResults)[0]
-      if (id) {
-        let info = this.elementResults[id];
-        if (info) {
-          for (let item in info.visibilityInfo) {
-            props.push(item);
-          }
-        }
-      }
-    }
-    return props;
-  }
-
-  // If one element information object has
-  // visibility information all
-  // element information objects will
-  getColorContrastProps () {
-    let props = [];
-    if (this.elementResults) {
-      let id = Object.keys(this.elementResults)[0]
-      if (id) {
-        let info = this.elementResults[id];
-        if (info) {
-          for (let item in info.ccrInfo) {
-            props.push(item);
-          }
-        }
-      }
-    }
-    return props;
-  }
-
-  // Check all element information objects,
-  // since not all elements use the same
-  // HTML attributes
-  getHTMLAttributeProps () {
-    let props = [];
-    for (let id in this.elementResults) {
-      let info = this.elementResults[id];
-      if (info.htmlAttrInfo) {
-        for (let item in info.htmlAttrInfo) {
-          if (props.indexOf(item) < 0) {
-            props.push(item);
-          }
-        }
-      }
-    }
-    return props;
-  }
-
-  // Check all element information objects,
-  // since not all elements use the same
-  // ARIA attributes
-  getARIAAttributeProps () {
-    let props = [];
-    for (let id in this.elementResults) {
-      let info = this.elementResults[id];
-      if (info.ariaAttrInfo) {
-        for (let item in info.ariaAttrInfo) {
-          if (props.indexOf(item) < 0) {
-            props.push(item);
-          }
-        }
-      }
-    }
-    return props;
-  }
-
-  addCSVColumnHeaders(props, flag) {
-    if (typeof flag !== 'boolean') {
-      flag = true;
-    }
-    let csv = '';
-    for (let i = 0; i < props.length; i += 1) {
-      if ((i !== 0) || flag) {
-        csv += ',';
-      }
-      csv += '"' + props[i] + '"';
-    }
-    console.log('[addSCVColumnHeader][csv]: ' + csv);
-    return csv;
   }
 
   toCSV () {
-
-    let accNameProps = this.getAccNameProps();
-    let ccrProps     = this.getColorContrastProps();
-    let visProps     = this.getVisibilityProps();
-    let htmlAttrProps    = this.getHTMLAttributeProps();
-    let ariaAttrProps    = this.getARIAAttributeProps();
-
-    let csv = '';
-    csv += this.addCSVColumnHeaders(basicProps, false);
-    csv += this.addCSVColumnHeaders(accNameProps);
-    csv += this.addCSVColumnHeaders(ccrProps);
-    csv += this.addCSVColumnHeaders(visProps);
-    csv += this.addCSVColumnHeaders(htmlAttrProps);
-    csv += this.addCSVColumnHeaders(ariaAttrProps);
-    csv += '\n';
-
-    for (let er in this.elementResults) {
-      let info = this.elementResults[er];
-      csv += this.resultElementInfo.toCSV(info, basicProps, accNameProps, ccrProps, visProps, htmlAttrProps, ariaAttrProps);
-    }
-    return csv
+    let viewCSV = new ViewRuleResultCSV(this.elementResults);
+    return viewCSV.getCSV();
   }
 
   resize (mainHeight) {
